DebianMed/requirements.txt
appdirs
backcall
black
boltons
CacheControl
certifi
chardet
click
colorama
contextlib2
decorator
distlib
distro
html5lib
idna
ipaddr
ipython
ipython-genutils
jedi
lockfile
msgpack
mypy-extensions
packaging
parso
pathspec
pep517
pexpect
pickleshare
progress
prompt-agentkit
psycopg2
ptyprocess
Pygments
pyparsing
pytoml
PyYAML
regex
requests
retrying
ruamel.yaml
ruamel.yaml.clib
six
toml
traitlets
typed-ast
typing-extensions
urllib3
wcwidth
webencodings
